Refactor: 임시저장 목록 페이지 loader 추가
